move this logic to it's own processor
