+++ lib/heaven.rb
@@ -40,3 +40,4 @@ require "heaven/jobs"
 require "heaven/provider"
 require "heaven/provisioner"
 require "heaven/notifier"
+require "heaven/pull_request_processor"
